+++ core_js/google_link_fix.js
@@ -25,8 +25,8 @@
     "use strict";
 
     function injectFunction() {
-        var ele = document.createElement('script');
-        var s = document.getElementsByTagName('script')[0];
+        let ele = document.createElement('script');
+        let s = document.getElementsByTagName('script')[0];
 
         ele.type = 'text/javascript';
         ele.textContent = "Object.defineProperty(window, 'rwt', {" +
@@ -46,13 +46,13 @@
         injectFunction();
 
         document.addEventListener('mouseover', function (event) {
-            var a = event.target, depth = 1;
+            let a = event.target, depth = 1;
 
-            while (a && a.tagName != 'A' && depth-- > 0) {
+            while (a && a.tagName !== 'A' && depth-- > 0) {
                 a = a.parentNode;
             }
 
-            if (a && a.tagName == 'A') {
+            if (a && a.tagName === 'A') {
                 try {
                     a.removeAttribute('data-cthref');
                     delete a.dataset.cthref;