+++ core_js/write_version.js
@@ -20,5 +20,5 @@
  * This file writes only the version into every page.
  * @return version
  */
-var version = browser.runtime.getManifest().version;
+const version = browser.runtime.getManifest().version;
 $('#version').text(version);